requirements.txt
django==3.2.15
Pillow==9.2.0
environs[django]==9.3.2
djangorestframework==3.13.1
psycopg2-binary==2.9.3
whitenoise==5.3.0
drf-yasg==1.21.7